+++ config/components/keys.js
@@ -1,8 +1,10 @@
 'use strict'
 
-const secret = '253ecb4dd4eeb36a943edd5c2510ef998866b0110596e44f'
 module.exports = {
   keys: {
-    secret: secret
+    secret: '253ecb4dd4eeb36a943edd5c2510ef998866b0110596e44f',
+    session: '7c213c2bf2591451713847e8724e01ce199b8d58009e3551',
+    githubId: 'c1435469b68887fa07b4',
+    githubSecret: '7b61102c2e2386e484f548d76c3dc6334df7dd95'
   }
 }