+++ core/index.js
@@ -1,5 +1,6 @@
 'use strict'
 
 const hash = require('./hash')
+const listDir = require('./listDir')
 
-module.exports = Object.assign({}, hash)
+module.exports = Object.assign({}, hash, listDir)