+++ server/routes/api/v1/user/get.js
@@ -2,18 +2,6 @@
 'use strict'
 
 const User = rootRequire('server/models/user')
-const logger = require('winston')
 
-module.exports = async (ctx, next) => {
-  let users
-
-  try {
-    users = await User.all()
-  } catch (err) {
-    logger.error('Failed to receive all users.' + err)
-    ctx.status = 500
-    ctx.body = 'Failed'
-  }
-
-  ctx.body = JSON.stringify(users)
-}
+module.exports = async (ctx, next) =>
+  (ctx.body = JSON.stringify(await User.all()))