+++ server/routes/login/post.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const passport = require('koa-passport')
+
+module.exports = passport.authenticate('local', {
+  successRedirect: '/dashboard',
+  failureRedirect: '/login',
+  failureFlash: true
+})